requirements.txt
# 🚀 GigShield Backend Requirements

fastapi
uvicorn
sqlalchemy
psycopg2-binary
requests
PyJWT
joblib
